fix netex tests and some refacto
